requirements.txt
flask==3.0.3
yfinance==0.2.43
pandas==2.2.2
numpy==1.26.4
plotly==5.24.1
requests==2.32.3

# MongoDB Atlas client
pymongo==4.8.0

# Load env vars from .env during development
python-dotenv==1.0.1

# Additional dependencies for MongoDB integration
matplotlib==3.8.2
seaborn==0.13.0

